requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.6.2
astroid==2.15.6
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==39.0.2
cycler==0.11.0
dill==0.3.7
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
et-xmlfile==1.1.0
fastapi==0.101.1
fonttools==4.42.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
numpy==1.25.1
openpyxl==3.1.2
orjson==3.9.1
packaging==23.1
pandas==2.0.3
passlib==1.7.4
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.7
psycopg2-binary==2.9.6
pyarrow==13.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
pydantic_core==2.4.0
pylance==0.7.3
pylint==2.17.5
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
response==0.5.0
rsa==4.9
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.0
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.18.1
websockets==10.4
wrapt==1.15.0
